add getbalance function
