+++ frontend/index.html
@@ -65,7 +65,7 @@
                 </li>
             </ul>
             <div class="balance">
-                <span>Баланс: </span><span class="text-primary">500 $</span>
+                <span>Баланс: </span><span class="text-primary" id="balance"></span>
             </div>
         </div>
 
@@ -77,7 +77,7 @@
                 </svg>
             </div>
             <div class="sidebar-user-name position-relative" data-bs-toggle="dropdown" aria-expanded="false">
-                <span id="sidebar-user-name">RomanChernov</span>
+                <span id="sidebar-user-name"></span>
                 <ul class="dropdown-menu dropdown-content text-small shadow position-absolute">
                     <li>
                         <a class="dropdown-item rounded" href="#">Sign out</a>
