+++ frontend/src/components/navbar.js
@@ -1,13 +1,18 @@
+import {Auth} from "../services/auth.js";
+import {CustomHttp} from "../services/custom-http.js";
+import config from "../../config/config.js";
+
 export class Navbar {
     constructor() {
         this.dropdownElement = document.getElementById('dropdown');
-        this.dropDownButton = document.getElementById('dropdownBtn');
         this.sidebarElement = document.getElementById('sidebar');
-        this.sectionElement = document.getElementById('section');
         this.sidebarToggleButton = document.getElementById('sidebarToggle');
         this.sidebarToggleButtonIcon = document.querySelector('#sidebarToggle i');
         this.listItems = document.querySelectorAll('#sidebar > .main-part > ul > li a');
         this.clickList = document.querySelectorAll('.click-btn');
+        this.userNameElement = document.getElementById('sidebar-user-name');
+        this.balanceElement = document.getElementById('balance');
+
 
         this.dropdownElement.addEventListener('click', () => {
             this.dropdownElement.classList.toggle('dropdown-active');
@@ -17,6 +22,12 @@ export class Navbar {
         this.navbarSizeChange();
         this.navbarClickClose();
 
+        const userInfo = Auth.getUserInfo();
+        const accessToken = localStorage.getItem(Auth.accessTokenKey);
+        if (userInfo && accessToken) {
+            this.userNameElement.innerHTML = `${userInfo.name} ${userInfo.lastName}`;
+        }
+        this.getBalance();
 
     }
 
@@ -58,6 +69,19 @@ export class Navbar {
                     this.sidebarElement.classList.add('collapsed');
                 })
             });
+        }
+    }
+
+    async getBalance() {
+        try {
+            const result = await CustomHttp.request(config.host + '/balance');
+            if (result.error) {
+                throw new Error(result.error)
+            }
+            console.log(result);
+            this.balanceElement.innerText = `${result.balance} $`;
+        } catch (error) {
+            console.log(error)
         }
     }
 }
