+++ frontend/src/router.js
@@ -9,7 +9,8 @@ export class Router {
         this.contentElement = document.getElementById('content');
         this.titleElement = document.getElementById('page-title');
         this.stylesElement = document.getElementById('styles');
-        this.userNameElement = document.getElementById('sidebar-user-name');
+
+
 
         this.routes = [
             {
@@ -129,19 +130,17 @@ export class Router {
         const newRoute = this.routes.find(item => {
             return item.route === urlRoute;
         });
+
         if (!newRoute) {
            return window.location.href = '#/signup';
         }
+
         this.stylesElement.setAttribute('href', newRoute.styles);
         this.contentElement.innerHTML =
             await fetch(newRoute.template).then(response => response.text());
         this.titleElement.innerText = newRoute.title;
 
-        const userInfo = Auth.getUserInfo();
-        const accessToken = localStorage.getItem(Auth.accessTokenKey);
-        if (userInfo && accessToken) {
-            this.userNameElement.innerHTML = `${userInfo.name} ${userInfo.lastName}`;
-        }
+
 
         newRoute.load();
